# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MegaOctav API",
        "description": "Dokumentasi API untuk sistem reservasi MegaOctav",
        "contact": {
            "email": "imeg290@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Ambil semua transaksi",
                "operationId": "ceac716339cb2455648899fac415b98f",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar transaksi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "id_user": {
                                                "type": "integer"
                                            },
                                            "schedule_id": {
                                                "type": "integer"
                                            },
                                            "payment_method": {
                                                "type": "string"
                                            },
                                            "total_price": {
                                                "type": "number"
                                            },
                                            "payment_date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Buat transaksi baru",
                "operationId": "d3a7a8d8ab4eeb1b4fb29af77c0e98f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_user",
                                    "schedule_id",
                                    "payment_method",
                                    "total_price",
                                    "payment_date"
                                ],
                                "properties": {
                                    "id_user": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "schedule_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "gopay"
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "example": 100000
                                    },
                                    "payment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-21"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaksi berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "id_user": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "schedule_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "payment_method": {
                                            "type": "string",
                                            "example": "gopay"
                                        },
                                        "total_price": {
                                            "type": "number",
                                            "example": 100000
                                        },
                                        "payment_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-05-21"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-22T10:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-22T10:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Lihat transaksi berdasarkan ID",
                "operationId": "a64a50b3720f06a3a57ca6d76fae5b12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail transaksi"
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Update transaksi",
                "operationId": "456e172c964e6149c8b721e6d735a6b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment_method": {
                                        "type": "string",
                                        "example": "qris"
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "example": 120000
                                    },
                                    "payment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-22"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaksi berhasil diupdate"
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Hapus transaksi",
                "operationId": "4c4238c89f4357bbeb6e0dce40b7a343",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Transaksi berhasil dihapus"
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan"
                    }
                }
            }
        },
...
    },
...
}